+++ AddOns/TeleportMenu/TeleportMenu.lua
@@ -19,109 +19,6 @@ local globalWidth, globalHeight = 40, 40 -- defaults
 -- Teleport Tables
 --------------------------------------
 
-local covenantsMaxed = nil
-local function GetCovenantData(id) -- the id is the achievement criteria index from Re-Re-Re-Renowned
-	if covenantsMaxed then
-		return covenantsMaxed[id]
-	end
-	covenantsMaxed = {}
-	for i = 1, 4 do
-		local _, _, completed = GetAchievementCriteriaInfo(15646, i)
-		covenantsMaxed[i] = completed
-	end
-end
-
-local availableHearthstones = {}
-local hearthstoneToys = {
-	[54452] = true, -- Ethereal Portal
-	[64488] = true, -- The Innkeeper's Daughter
-	[93672] = true, -- Dark Portal
-	[142542] = true, -- Tome of Town Portal
-	[162973] = true, -- Greatfather Winter's Hearthstone
-	[163045] = true, -- Headless Horseman's Hearthstone
-	[163206] = true, -- Weary Spirit Binding
-	[165669] = true, -- Lunar Elder's Hearthstone
-	[165670] = true, -- Peddlefeet's Lovely Hearthstone
-	[165802] = true, -- Noble Gardener's Hearthstone
-	[166746] = true, -- Fire Eater's Hearthstone
-	[166747] = true, -- Brewfest Reveler's Hearthstone
-	[168907] = true, -- Holographic Digitalization Hearthstone
-	[172179] = true, -- Eternal Traveler's Hearthstone
-	[180290] = function()
-		-- Night Fae Hearthstone
-		if GetCovenantData(3) then
-			return true
-		end
-		local covenantID = C_Covenants.GetActiveCovenantID()
-		if covenantID == 3 then
-			return true
-		end
-	end,
-	[182773] = function()
-		-- Necrolord Hearthstone
-		if GetCovenantData(2) then
-			return true
-		end
-		local covenantID = C_Covenants.GetActiveCovenantID()
-		if covenantID == 4 then
-			return true
-		end
-	end,
-	[183716] = function()
-		-- Venthyr Sinstone
-		if GetCovenantData(4) then
-			return true
-		end
-		local covenantID = C_Covenants.GetActiveCovenantID()
-		if covenantID == 2 then
-			return true
-		end
-	end,
-	[184353] = function()
-		-- Kyrian Hearthstone
-		if GetCovenantData(1) then
-			return true
-		end
-		local covenantID = C_Covenants.GetActiveCovenantID()
-		if covenantID == 1 then
-			return true
-		end
-	end,
-	[188952] = true, -- Dominated Hearthstone
-	[190196] = true, -- Enlightened Hearthstone
-	[190237] = true, -- Broker Translocation Matrix
-	[193588] = true, -- Timewalker's Hearthstone
-	[200630] = true, -- Ohnir Windsage's Hearthstone
-	[206195] = true, -- Path of the Naaru
-	[208704] = true, -- Deepdweller's Earthen Hearthstone
-	[209035] = true, -- Hearthstone of the Flame
-	[210455] = function()
-		-- Draenic Hologem (Draenei and Lightforged Draenei only)
-		local _, _, raceId = UnitRace("player")
-		if raceId == 11 or raceId == 30 then
-			return true
-		end
-	end,
-	[212337] = true, -- Stone of the Hearth
-	[228940] = true -- Notorious Thread's Hearthstone
-}
-
-local availableWormholes = {}
-local wormholes = {
-	30542, -- Dimensional Ripper - Area 52
-	18984, -- Dimensional Ripper - Everlook
-	18986, -- Ultrasafe Transporter: Gadgetzan
-	30544, -- Ultrasafe Transporter: Toshley's Station
-	48933, -- Wormhole Generator: Northrend
-	87215, -- Wormhole Generator: Pandaria
-	112059, -- Wormhole Centrifuge (Dreanor) 6
-	151652, -- Wormhole Generator: Argus
-	168807, -- Wormhole Generator: Kul Tiras 5
-	168808, -- Wormhole Generator: Zandalar
-	172924, -- Wormhole Generator: Shadowlands 3
-	198156, -- Wyrmhole Generator: Dragon Isles 4
-	221966 -- Wormhole Generator: Khaz Algar
-}
 local availableSeasonalTeleports = {}
 
 local shortNames = {
@@ -163,6 +60,8 @@ local shortNames = {
 	[424187] = L["Atal'Dazar"],
 	[445418] = L["Siege of Boralus"],
 	[464256] = L["Siege of Boralus"],
+	[467553] = L["The MOTHERLODE!!"],
+	[467555] = L["The MOTHERLODE!!"],
 	-- SL
 	[354462] = L["The Necrotic Wake"],
 	[354463] = L["Plaguefall"],
@@ -200,6 +99,7 @@ local shortNames = {
 	[445444] = L["Priory of the Sacred Flame"],
 	[445417] = L["Ara-Kara, City of Echoes"],
 	[445441] = L["Darkflame Cleft"],
+	[1216786] = L["Operation: Floodgate"],
 	-- Mage teleports
 	[3561] = L["Stormwind"],
 	[3562] = L["Ironforge"],
@@ -254,37 +154,45 @@ local shortNames = {
 	[281402] = L["Dazar'alor"],
 	[344597] = L["Oribos"],
 	[395289] = L["Valdrakken"],
-	[446534] = L["Dornogal"]
+	[446534] = L["Dornogal"],
 }
 
 local tpTable = {
 	-- Hearthstones
-	{id = 6948, type = "item", hearthstone = true}, -- Hearthstone
-	{id = 556, type = "spell"}, -- Astral Recall (Shaman)
-	{id = 110560, type = "toy", quest = {34378, 34586}}, -- Garrison Hearthstone
-	{id = 140192, type = "toy", quest = {44184, 44663}}, -- Dalaran Hearthstone
+	{ id = 6948, type = "item", hearthstone = true }, -- Hearthstone
+	{ id = 556, type = "spell" }, -- Astral Recall (Shaman)
+	{ id = 110560, type = "toy", quest = { 34378, 34586 } }, -- Garrison Hearthstone
+	{ id = 140192, type = "toy", quest = { 44184, 44663 } }, -- Dalaran Hearthstone
 	-- Engineering
-	{type = "wormholes", iconId = 4620673}, -- Engineering Wormholes
+	{ type = "wormholes", iconId = 4620673 }, -- Engineering Wormholes
+	{ type = "item_teleports", iconId = 133655 }, -- Item Teleports
 	-- Class Teleports
-	{id = 1, type = "flyout", iconId = 237509, subtype = "mage"}, -- Teleport (Mage) (Horde)
-	{id = 8, type = "flyout", iconId = 237509, subtype = "mage"}, -- Teleport (Mage) (Alliance)
-	{id = 11, type = "flyout", iconId = 135744, subtype = "mage"}, -- Portals (Mage) (Horde)
-	{id = 12, type = "flyout", iconId = 135748, subtype = "mage"}, -- Portals (Mage) (Alliance)
-	{id = 126892, type = "spell"}, -- Zen Pilgrimage (Monk)
-	{id = 50977, type = "spell"}, -- Death Gate (Death Knight)
-	{id = 193753, type = "spell"}, -- Dreamwalk (Druid)
+	{ id = 1, type = "flyout", iconId = 237509, subtype = "mage" }, -- Teleport (Mage) (Horde)
+	{ id = 8, type = "flyout", iconId = 237509, subtype = "mage" }, -- Teleport (Mage) (Alliance)
+	{ id = 11, type = "flyout", iconId = 135744, subtype = "mage" }, -- Portals (Mage) (Horde)
+	{ id = 12, type = "flyout", iconId = 135748, subtype = "mage" }, -- Portals (Mage) (Alliance)
+	{ id = 126892, type = "spell" }, -- Zen Pilgrimage (Monk)
+	{ id = 50977, type = "spell" }, -- Death Gate (Death Knight)
+	{ id = 18960, type = "spell" }, -- Teleport: Moonglade (Druid)
+	{ id = 193753, type = "spell" }, -- Dreamwalk (Druid) (replaces Teleport: Moonglade)
+	-- Racials
+	{ id = 312370, type = "spell" }, -- Make Camp (Vulpera)
+	{ id = 312372, type = "spell" }, -- Return to Camp (Vulpera)
+
 	-- Dungeon/Raid Teleports
-	{id = 230, type = "flyout", iconId = 574788, name = L["Cataclysm"], subtype = "path"}, -- Hero's Path: Cataclysm
-	{id = 84, type = "flyout", iconId = 328269, name = L["Mists of Pandaria"], subtype = "path"}, -- Hero's Path: Mists of Pandaria
-	{id = 96, type = "flyout", iconId = 1413856, name = L["Warlords of Draenor"], subtype = "path"}, -- Hero's Path: Warlords of Draenor
-	{id = 224, type = "flyout", iconId = 1260827, name = L["Legion"], subtype = "path"}, -- Hero's Path: Legion
-	{id = 223, type = "flyout", iconId = 1869493, name = L["Battle for Azeroth"], subtype = "path"}, -- Hero's Path: Battle for Azeroth
-	{id = 220, type = "flyout", iconId = 236798, name = L["Shadowlands"], subtype = "path"}, -- Hero's Path: Shadowlands
-	{id = 222, type = "flyout", iconId = 4062765, name = L["Shadowlands Raids"], subtype = "path"}, -- Hero's Path: Shadowlands Raids
-	{id = 227, type = "flyout", iconId = 4640496, name = L["Dragonflight"], subtype = "path"}, -- Hero's Path: Dragonflight
-	{id = 231, type = "flyout", iconId = 5342925, name = L["Dragonflight Raids"], subtype = "path"}, -- Hero's Path: Dragonflight Raids
-	{id = 232, type = "flyout", iconId = 5872031, name = L["The War Within"], subtype = "path"} -- Hero's Path: The War Within
+	{ id = 230, type = "flyout", iconId = 574788, name = L["Cataclysm"], subtype = "path" }, -- Hero's Path: Cataclysm
+	{ id = 84, type = "flyout", iconId = 328269, name = L["Mists of Pandaria"], subtype = "path" }, -- Hero's Path: Mists of Pandaria
+	{ id = 96, type = "flyout", iconId = 1413856, name = L["Warlords of Draenor"], subtype = "path" }, -- Hero's Path: Warlords of Draenor
+	{ id = 224, type = "flyout", iconId = 1260827, name = L["Legion"], subtype = "path" }, -- Hero's Path: Legion
+	{ id = 223, type = "flyout", iconId = 1869493, name = L["Battle for Azeroth"], subtype = "path" }, -- Hero's Path: Battle for Azeroth
+	{ id = 220, type = "flyout", iconId = 236798, name = L["Shadowlands"], subtype = "path" }, -- Hero's Path: Shadowlands
+	{ id = 222, type = "flyout", iconId = 4062765, name = L["Shadowlands Raids"], subtype = "path" }, -- Hero's Path: Shadowlands Raids
+	{ id = 227, type = "flyout", iconId = 4640496, name = L["Dragonflight"], subtype = "path" }, -- Hero's Path: Dragonflight
+	{ id = 231, type = "flyout", iconId = 5342925, name = L["Dragonflight Raids"], subtype = "path" }, -- Hero's Path: Dragonflight Raids
+	{ id = 232, type = "flyout", iconId = 5872031, name = L["The War Within"], subtype = "path" }, -- Hero's Path: The War Within
 }
+
+local GetItemCount = C_Item.GetItemCount
 
 --------------------------------------
 -- Texture Stuff
@@ -293,12 +201,10 @@ local tpTable = {
 local function SetTextureByItemId(frame, itemId)
 	frame:SetNormalTexture(DEFAULT_ICON) -- Temp while loading
 	local item = Item:CreateFromItemID(tonumber(itemId))
-	item:ContinueOnItemLoad(
-		function()
-			local icon = item:GetItemIcon()
-			frame:SetNormalTexture(icon)
-		end
-	)
+	item:ContinueOnItemLoad(function()
+		local icon = item:GetItemIcon()
+		frame:SetNormalTexture(icon)
+	end)
 end
 
 local function retrySetNormalTexture(button, itemId, attempt)
@@ -309,12 +215,9 @@ local function retrySetNormalTexture(button, itemId, attempt)
 		return
 	end
 	if attempts < 5 then
-		C_Timer.After(
-			1,
-			function()
-				retrySetNormalTexture(button, itemId, attempts + 1)
-			end
-		)
+		C_Timer.After(1, function()
+			retrySetNormalTexture(button, itemId, attempts + 1)
+		end)
 	else
 		print(APPEND .. L["Missing Texture %s"]:format(itemId))
 	end
@@ -324,12 +227,9 @@ local function retryGetToyTexture(toyId, attempt)
 	local attempts = attempt or 1
 	local _, name, texture = C_ToyBox.GetToyInfo(toyId)
 	if attempts < 5 then
-		C_Timer.After(
-			0.1,
-			function()
-				retryGetToyTexture(toyId, attempts + 1)
-			end
-		)
+		C_Timer.After(0.1, function()
+			retryGetToyTexture(toyId, attempts + 1)
+		end)
 	end
 end
 
@@ -349,13 +249,15 @@ local function setToolTip(self, type, id, hs)
 	GameTooltip:SetOwner(self, "ANCHOR_NONE")
 	local yOffset = globalHeight / 2
 	GameTooltip:SetPoint("BOTTOMLEFT", TeleportMeButtonsFrame, "TOPRIGHT", 0, yOffset)
-	if hs and db.hearthstone and db.hearthstone == "rng" then
+	if hs and db["Teleports:Hearthstone"] and db["Teleports:Hearthstone"] == "rng" then
 		local bindLocation = GetBindLocation()
 		GameTooltip:SetText(L["Random Hearthstone"], 1, 1, 1)
 		GameTooltip:AddLine(L["Random Hearthstone Tooltip"], 1, 1, 1)
 		GameTooltip:AddLine(L["Random Hearthstone Location"]:format(bindLocation), 1, 1, 1)
 	elseif type == "item" then
 		GameTooltip:SetItemByID(id)
+	elseif type == "item_teleports" then
+		GameTooltip:SetText(L["Item Teleports"] .. "\n" .. L["Item Teleports Tooltip"], 1, 1, 1)
 	elseif type == "toy" then
 		GameTooltip:SetToyByItemID(id)
 	elseif type == "spell" then
@@ -369,7 +271,10 @@ local function setToolTip(self, type, id, hs)
 			GameTooltip:SetText(professionInfo.professionName, 1, 1, 1)
 		end
 	elseif type == "seasonalteleport" then
-		GameTooltip:SetText(L["Seasonal Teleports"], 1, 1, 1)
+		local currExpID = GetExpansionLevel()
+		local expName = _G["EXPANSION_NAME" .. currExpID]
+		local title = MYTHIC_DUNGEON_SEASON:format(expName, tpm.settings.current_season)
+		GameTooltip:SetText(title, 1, 1, 1)
 		GameTooltip:AddLine(L["Seasonal Teleports Tooltip"], 1, 1, 1)
 	end
 	GameTooltip:Show()
@@ -438,6 +343,7 @@ local function createFlyOutButton(flyOutFrame, flyoutData, tooltipData) -- Flyou
 	function flyOutButton:SetFlyOutFrame(frame)
 		flyOutButton.flyoutFrame = frame
 	end
+
 	flyOutButton:SetFlyOutFrame(flyOutFrame)
 
 	function flyOutButton:Recycle()
@@ -458,30 +364,24 @@ local function createFlyOutButton(flyOutFrame, flyoutData, tooltipData) -- Flyou
 		tooltipType = tooltipData.type
 		tooltipId = tooltipData.id
 	end
-	flyOutButton:SetScript(
-		"OnEnter",
-		function(self)
-			if InCombatLockdown() then
-				setCombatTooltip(self)
-				return
-			end
-			CloseAllFlyouts()
-			setToolTip(self, tooltipType, tooltipId)
-			self.flyoutFrame:Show()
+	flyOutButton:SetScript("OnEnter", function(self)
+		if InCombatLockdown() then
+			setCombatTooltip(self)
+			return
 		end
-	)
-	flyOutButton:SetScript(
-		"OnLeave",
-		function(self)
-			GameTooltip:Hide()
-		end
-	)
+		CloseAllFlyouts()
+		setToolTip(self, tooltipType, tooltipId)
+		self.flyoutFrame:Show()
+	end)
+	flyOutButton:SetScript("OnLeave", function(self)
+		GameTooltip:Hide()
+	end)
 
 	-- Text
-	flyOutButton.text:SetFont(STANDARD_TEXT_FONT, 13, "OUTLINE")
+	flyOutButton.text:SetFont(STANDARD_TEXT_FONT, db["Button:Text:Size"], "OUTLINE")
 	flyOutButton.text:SetTextColor(1, 1, 1, 1)
 	flyOutButton.text:Hide()
-	if db.buttonText == true and flyoutData.name then
+	if db["Button:Text:Show"] == true and flyoutData.name then
 		flyOutButton.text:SetText(flyoutData.name)
 		flyOutButton.text:Show()
 	end
@@ -517,26 +417,23 @@ local function createFlyOutFrame()
 	flyOutFrame:SetFrameLevel(103)
 	flyOutFrame:SetPropagateMouseClicks(true)
 	flyOutFrame:SetPropagateMouseMotion(true)
-	flyOutFrame:SetScript(
-		"OnLeave",
-		function(self)
-			GameTooltip:Hide()
-			if not InCombatLockdown() then -- XXX Needed?
-				self:Hide()
-			end
+	flyOutFrame:SetScript("OnLeave", function(self)
+		GameTooltip:Hide()
+		if not InCombatLockdown() then -- XXX Needed?
+			self:Hide()
 		end
-	)
+	end)
 
 	flyOutFrame:Hide()
 	return flyOutFrame
 end
 
--- Args
--- frame: Parent Frame
--- type: item, spell, toy type for the button click
--- text: Text to display on the button
--- id: id of the item, spell, or toy
--- hearthstone: boolean if the button is for a hearthstone (only used for tooltip atm)
+---@param frame Frame
+---@param type string
+---@param text string|nil
+---@param id integer
+---@param hearthstone? boolean
+---@return Frame
 local function CreateSecureButton(frame, type, text, id, hearthstone)
 	local button
 	if next(secureButtonsPool) then
@@ -561,33 +458,24 @@ local function CreateSecureButton(frame, type, text, id, hearthstone)
 	button:RegisterForClicks("AnyDown", "AnyUp")
 
 	-- Text
-	button.text:SetFont(STANDARD_TEXT_FONT, 13, "OUTLINE")
+	button.text:SetFont(STANDARD_TEXT_FONT, db["Button:Text:Size"], "OUTLINE")
 	button.text:SetTextColor(1, 1, 1, 1)
 	button.text:Hide()
-	if db.buttonText == true and text then
+	if db["Button:Text:Show"] == true and text then
 		button.text:SetText(text)
 		button.text:Show()
 	end
 
 	-- Scripts
-	button:SetScript(
-		"OnLeave",
-		function(self)
-			GameTooltip:Hide()
-		end
-	)
-	button:SetScript(
-		"OnEnter",
-		function(self)
-			setToolTip(self, type, id, hearthstone)
-		end
-	)
-	button:SetScript(
-		"OnShow",
-		function(self)
-			self.cooldownFrame:CheckCooldown(id, type)
-		end
-	)
+	button:SetScript("OnLeave", function(self)
+		GameTooltip:Hide()
+	end)
+	button:SetScript("OnEnter", function(self)
+		setToolTip(self, type, id, hearthstone)
+	end)
+	button:SetScript("OnShow", function(self)
+		self.cooldownFrame:CheckCooldown(id, type)
+	end)
 	button.cooldownFrame:CheckCooldown(id, type)
 
 	-- Textures
@@ -625,71 +513,49 @@ function tpm:GetIconText(spellId)
 	if text then
 		return text
 	end
-	print(APPEND .. "No short name found for spellID " .. id .. ", please report this on GitHub")
+	print(APPEND .. "No short name found for spellID " .. spellId .. ", please report this on GitHub")
 end
 
-function tpm:GetAvailableHearthstoneToys()
-	local hearthstoneNames = {}
-	for _, toyId in pairs(availableHearthstones) do
-		local _, name, texture = C_ToyBox.GetToyInfo(toyId)
-		if not texture then
-			texture = DEFAULT_ICON
-		end
-		if not name then
-			name = tostring(toyId)
-		end
-		hearthstoneNames[toyId] = {name = name, texture = texture}
-	end
-	return hearthstoneNames
-end
-
-function tpm:updateAvailableHearthstones()
-	availableHearthstones = {}
-	for id, usable in pairs(hearthstoneToys) do
-		if PlayerHasToy(id) then
-			if type(usable) == "function" and usable() then
-				table.insert(availableHearthstones, id)
-			elseif usable == true then
-				table.insert(availableHearthstones, id)
-			end
-		end
-	end
-end
-
-function tpm:updateAvailableWormholes()
-	for _, id in ipairs(wormholes) do
-		if PlayerHasToy(id) and C_ToyBox.IsToyUsable(id) then
-			table.insert(availableWormholes, id)
-		end
-	end
-end
-
-function tpm:updateAvailableSeasonalTeleport()
+function tpm:UpdateAvailableSeasonalTeleports()
+	availableSeasonalTeleports = {}
 	local playerFaction = UnitFactionGroup("player")
 	local siegeOfBoralus = -1
+	local motherlode = -1
 	if playerFaction == "Alliance" then
 		siegeOfBoralus = 445418
+		motherlode = 467553
 	else
 		siegeOfBoralus = 464256
+		motherlode = 467555
 	end
 
-	local challengeMapIdTospellID = {
-		[353] = siegeOfBoralus, -- Siege of Boralus has two spells one for alliance and one for horde
-		[375] = 354464, -- Mists
-		[376] = 354462, -- Necrotic Wake
-		[499] = 445444, -- Priory
-		[500] = 445443, -- The Rookery
-		[501] = 445269, -- Stonevault
-		[502] = 445416, -- City of Threads
-		[503] = 445417, -- Ara Ara
-		[504] = 445441, -- Darkflame Cleft
-		[505] = 445414, -- The Dawnbreaker
-		[506] = 445440, -- Cinderbrew Meadery
-		[507] = 445424 -- Grim Batol
+	local seasonalTeleports = {
+		-- TWW S1
+		[1] = {
+			[353] = siegeOfBoralus, -- Siege of Boralus has two spells one for alliance and one for horde
+			[375] = 354464, -- Mists
+			[376] = 354462, -- Necrotic Wake
+			[501] = 445269, -- Stonevault
+			[502] = 445416, -- City of Threads
+			[503] = 445417, -- Ara Ara
+			[505] = 445414, -- The Dawnbreaker
+			[507] = 445424, -- Grim Batol
+		},
+		-- TWW S2
+		[2] = {
+			[247] = motherlode, -- The MOTHERLODE!!
+			[370] = 373274, -- Operation: Mechagon - Workshop
+			[382] = 354467, -- Theater of Pain
+			[499] = 445444, -- Priory of the Sacred Flame
+			[500] = 445443, -- The Rookery
+			[504] = 445441, -- Darkflame Cleft
+			[506] = 445440, -- Cinderbrew Meadery
+			[525] = 1216786, -- Operation: Floodgate
+		},
 	}
 
 	for _, mapId in ipairs(C_ChallengeMode.GetMapTable()) do
-		local spellID = challengeMapIdTospellID[mapId]
+		local spellID = seasonalTeleports[tpm.settings.current_season][mapId]
 		if spellID and IsSpellKnown(spellID) then
 			table.insert(availableSeasonalTeleports, spellID)
 		end
@@ -709,7 +575,7 @@ function tpm:checkQuestCompletion(quest)
 end
 
 function tpm:CreateFlyout(flyoutData)
-	if db.showOnlySeasonalHerosPath and flyoutData.subtype == "path" then
+	if db["Teleports:Seasonal:Only"] and flyoutData.subtype == "path" then
 		return
 	end
 	local _, _, spells, flyoutKnown = GetFlyoutInfo(flyoutData.id)
@@ -729,7 +595,7 @@ function tpm:CreateFlyout(flyoutData)
 	local flyoutsCreated = 0
 	local rowNr = 1
 
-	local inverse = db.reverseMageFlyouts and flyoutData.subtype == "mage"
+	local inverse = db["Teleports:Mage:Reverse"] and flyoutData.subtype == "mage"
 	local start, endLoop, step = 1, spells, 1
 	if inverse then -- Inverse loop params
 		start, endLoop, step = spells, 1, -1
@@ -737,7 +603,7 @@ function tpm:CreateFlyout(flyoutData)
 	for i = start, endLoop, step do
 		local spellId = select(1, GetFlyoutSlotInfo(flyoutData.id, i))
 		if IsSpellKnown(spellId) then
-			if flyoutsCreated == db.maxFlyoutIcons then
+			if flyoutsCreated == db["Flyout:Max_Per_Row"] then
 				flyoutsCreated = 0
 				rowNr = rowNr + 1
 			end
@@ -748,7 +614,7 @@ function tpm:CreateFlyout(flyoutData)
 		end
 	end
 
-	local frameWidth = rowNr > 1 and globalWidth * (db.maxFlyoutIcons + 1) or globalWidth * (flyoutsCreated + 1)
+	local frameWidth = rowNr > 1 and globalWidth * (db["Flyout:Max_Per_Row"] + 1) or globalWidth * (flyoutsCreated + 1)
 	flyOutFrame:SetSize(frameWidth, globalHeight * rowNr)
 	button.childButtons = childButtons
 	return button
@@ -759,8 +625,8 @@ function tpm:CreateSeasonalTeleportFlyout()
 		return
 	end
 
-	local tooltipData = {type = "seasonalteleport"}
-	local seasonalFlyOutData = {id = -1, name = L["Season 1"], iconId = 5927657}
+	local tooltipData = { type = "seasonalteleport" }
+	local seasonalFlyOutData = { id = -1, name = L["Season " .. tpm.settings.current_season], iconId = 5927657 }
 	local yOffset = -globalHeight * TeleportMeButtonsFrame:GetButtonAmount()
 
 	local flyOutFrame = createFlyOutFrame()
@@ -786,7 +652,7 @@ function tpm:CreateSeasonalTeleportFlyout()
 end
 
 function tpm:CreateWormholeFlyout(flyoutData)
-	if #availableWormholes == 0 then
+	if #tpm.AvailableWormholes == 0 then
 		return
 	end
 
@@ -795,13 +661,38 @@ function tpm:CreateWormholeFlyout(flyoutData)
 	local flyOutFrame = createFlyOutFrame()
 	flyOutFrame:SetPoint("LEFT", TeleportMeButtonsFrame, "TOPRIGHT", 0, yOffset)
 
-	local button = createFlyOutButton(flyOutFrame, flyoutData, {type = "profession", id = 202})
+	local button = createFlyOutButton(flyOutFrame, flyoutData, { type = "profession", id = 202 })
 	button:SetPoint("LEFT", TeleportMeButtonsFrame, "TOPRIGHT", 0, yOffset)
 
 	local flyoutsCreated = 0
-	for _, wormholeId in ipairs(availableWormholes) do
+	for _, wormholeId in ipairs(tpm.AvailableWormholes) do
 		flyoutsCreated = flyoutsCreated + 1
 		local flyOutButton = CreateSecureButton(flyOutFrame, "toy", nil, wormholeId)
+		local xOffset = globalWidth * flyoutsCreated
+		flyOutButton:SetPoint("TOPLEFT", flyOutFrame, "TOPLEFT", xOffset, 0)
+	end
+	flyOutFrame:SetSize(globalWidth * (flyoutsCreated + 1), globalHeight)
+
+	return button
+end
+
+function tpm:CreateItemTeleportsFlyout(flyoutData)
+	if #tpm.AvailableItemTeleports == 0 then
+		return
+	end
+
+	local yOffset = -globalHeight * TeleportMeButtonsFrame:GetButtonAmount()
+
+	local flyOutFrame = createFlyOutFrame()
+	flyOutFrame:SetPoint("LEFT", TeleportMeButtonsFrame, "TOPRIGHT", 0, yOffset)
+
+	local button = createFlyOutButton(flyOutFrame, flyoutData, { type = "item_teleports" })
+	button:SetPoint("LEFT", TeleportMeButtonsFrame, "TOPRIGHT", 0, yOffset)
+
+	local flyoutsCreated = 0
+	for _, itemTeleportId in ipairs(tpm.AvailableItemTeleports) do
+		flyoutsCreated = flyoutsCreated + 1
+		local flyOutButton = CreateSecureButton(flyOutFrame, "item", nil, itemTeleportId)
 		local xOffset = globalWidth * flyoutsCreated
 		flyOutButton:SetPoint("TOPLEFT", flyOutFrame, "TOPLEFT", xOffset, 0)
 	end
@@ -815,22 +706,19 @@ function tpm:updateHearthstone()
 	if not hearthstoneButton then
 		return
 	end
-	local texture
-	if db.hearthstone == "rng" then
-		local rng = math.random(#availableHearthstones)
+
+	if db["Teleports:Hearthstone"] == "rng" then
+		local rng = math.random(#tpm.AvailableHearthstones)
 		hearthstoneButton:SetNormalTexture(1669494) -- misc_rune_pvp_random
 		hearthstoneButton:SetAttribute("type", "toy")
-		hearthstoneButton:SetAttribute("toy", availableHearthstones[rng])
-	elseif db.hearthstone ~= "none" then
-		SetTextureByItemId(hearthstoneButton, db.hearthstone)
+		hearthstoneButton:SetAttribute("toy", tpm.AvailableHearthstones[rng])
+	elseif db["Teleports:Hearthstone"] ~= "none" then
+		SetTextureByItemId(hearthstoneButton, db["Teleports:Hearthstone"])
 		hearthstoneButton:SetAttribute("type", "toy")
-		hearthstoneButton:SetAttribute("toy", db.hearthstone)
-		hearthstoneButton:SetScript(
-			"OnEnter",
-			function(self)
-				setToolTip(self, "toy", db.hearthstone, true)
-			end
-		)
+		hearthstoneButton:SetAttribute("toy", db["Teleports:Hearthstone"])
+		hearthstoneButton:SetScript("OnEnter", function(self)
+			setToolTip(self, "toy", db["Teleports:Hearthstone"], true)
+		end)
 	else
 		if C_Item.GetItemCount(6948) == 0 then
 			print(APPEND .. L["No Hearthtone In Bags"])
@@ -840,48 +728,28 @@ function tpm:updateHearthstone()
 		SetTextureByItemId(hearthstoneButton, 6948)
 		hearthstoneButton:SetAttribute("type", "item")
 		hearthstoneButton:SetAttribute("item", "item:6948")
-		hearthstoneButton:SetScript(
-			"OnEnter",
-			function(self)
-				setToolTip(self, "item", 6948, true)
-			end
-		)
+		hearthstoneButton:SetScript("OnEnter", function(self)
+			setToolTip(self, "item", 6948, true)
+		end)
 	end
 	hearthstoneButton:Show()
-end
-
-function tpm:GetRandomHearthstone(retry)
-	if #availableHearthstones == 0 then
-		return
-	end
-	if #availableHearthstones == 1 then
-		return availableHearthstones[1]
-	end -- Don't even bother
-	local randomHs = availableHearthstones[math.random(#availableHearthstones)]
-	if lastRandomHearthstone == randomHs then -- Don't fully randomize, always a new one
-		randomHs = self:GetRandomHearthstone(true)
-	end
-	if not retry then
-		lastRandomHearthstone = randomHs
-	end
-	return randomHs
 end
 
 local function createAnchors()
 	if InCombatLockdown() then
 		return
 	elseif TeleportMeButtonsFrame and not TeleportMeButtonsFrame.reload then
-		if not db.enabled then
+		if not db["Enabled"] then
 			TeleportMeButtonsFrame:Hide()
 			return
 		end
-		if TeleportMeButtonsFrame:IsVisible() and db.hearthstone and db.hearthstone == "rng" then
+		if TeleportMeButtonsFrame:IsVisible() and db["Teleports:Hearthstone"] and db["Teleports:Hearthstone"] == "rng" then
 			local rng = tpm:GetRandomHearthstone()
 			TeleportMeButtonsFrame.hearthstoneButton:SetAttribute("toy", rng)
 		end
 		return
 	end
-	if not db.enabled then
+	if not db["Enabled"] then
 		return
 	end
 	local buttonsFrame = TeleportMeButtonsFrame or CreateFrame("Frame", "TeleportMeButtonsFrame", GameMenuFrame)
@@ -904,26 +772,30 @@ local function createAnchors()
 		local known
 
 		-- Checks and overwrites
-		if teleport.hearthstone and db.hearthstone ~= "none" then -- Overwrite main HS with user set HS
+		if teleport.hearthstone and db["Teleports:Hearthstone"] ~= "none" then -- Overwrite main HS with user set HS
 			tpm:DebugPrint("Overwriting main HS with user set HS")
 			teleport.type = "toy"
 			known = true
-			if db.hearthstone == "rng" then
+			if db["Teleports:Hearthstone"] == "rng" then
 				texture = 1669494 -- misc_rune_pvp_random
 				teleport.id = tpm:GetRandomHearthstone()
 			else
-				teleport.id = db.hearthstone
+				teleport.id = db["Teleports:Hearthstone"]
 			end
 			tpm:DebugPrint("Overwrite Info:", known, teleport.id, teleport.type, texture)
 		elseif teleport.type == "item" and C_Item.GetItemCount(teleport.id) > 0 then
 			known = true
-		elseif teleport.type == "toy" and PlayerHasToy(teleport.id) then
+		elseif
+			teleport.type == "toy" and PlayerHasToy(teleport.id --[[@as integer]])
+		then
 			if teleport.quest then
 				known = tpm:checkQuestCompletion(teleport.quest)
 			else
 				known = true
 			end
-		elseif teleport.type == "spell" and IsSpellKnown(teleport.id) then
+		elseif
+			teleport.type == "spell" and IsSpellKnown(teleport.id --[[@as integer]])
+		then
 			known = true
 		end
 
@@ -934,7 +806,7 @@ local function createAnchors()
 		-- Create Stuff
 		if known and (teleport.type == "toy" or teleport.type == "item" or teleport.type == "spell") then
 			tpm:DebugPrint(teleport.hearthstone)
-			local button = CreateSecureButton(buttonsFrame, teleport.type, nil, teleport.id, teleport.hearthstone)
+			local button = CreateSecureButton(buttonsFrame, teleport.type, nil, teleport.id --[[@as integer]], teleport.hearthstone)
 			local yOffset = -globalHeight * buttonsFrame:GetButtonAmount()
 			button:SetPoint("LEFT", buttonsFrame, "TOPRIGHT", 0, yOffset)
 			if teleport.hearthstone then -- store to replace item later
@@ -943,6 +815,11 @@ local function createAnchors()
 			buttonsFrame:IncrementButtons()
 		elseif teleport.type == "wormholes" then
 			local created = tpm:CreateWormholeFlyout(teleport)
+			if created then
+				buttonsFrame:IncrementButtons()
+			end
+		elseif teleport.type == "item_teleports" then
+			local created = tpm:CreateItemTeleportsFlyout(teleport)
 			if created then
 				buttonsFrame:IncrementButtons()
 			end
@@ -966,9 +843,9 @@ local function createAnchors()
 end
 
 function tpm:ReloadFrames()
-	if db.iconSize then
-		globalWidth = db.iconSize
-		globalHeight = db.iconSize
+	if db["Button:Size"] then
+		globalWidth = db["Button:Size"]
+		globalHeight = db["Button:Size"]
 	end
 
 	for _, button in ipairs(flyOutButtons) do
@@ -990,10 +867,17 @@ end
 SLASH_TPMENU1 = "/tpm"
 SLASH_TPMENU2 = "/tpmenu"
 SlashCmdList["TPMENU"] = function(msg)
-	print(APPEND .. L["Opening Options Menu"])
-	Settings.OpenToCategory(tpm:GetOptionsCategory())
-end
+	local args = { (" "):split(msg:lower()) }
+	msg = args[1]
 
+	if msg == "" then
+		Settings.OpenToCategory(tpm:GetOptionsCategory())
+	elseif msg == "filters" then
+		Settings.OpenToCategory(tpm:GetOptionsCategory(msg))
+	else
+		print(APPEND .. " unknown command: " .. msg)
+	end
+end
 --------------------------------------
 -- Loading
 --------------------------------------
@@ -1004,20 +888,23 @@ local function checkItemsLoaded(self)
 	end
 
 	self.continuableContainer = ContinuableContainer:Create()
+
 	local function LoadItems(itemTable)
-		for _, itemId in ipairs(itemTable) do
-			self.continuableContainer:AddContinuable(Item:CreateFromItemID(tonumber(itemId)))
+		for id, _ in ipairs(itemTable) do
+			self.continuableContainer:AddContinuable(Item:CreateFromItemID(id))
 		end
 	end
 
-	LoadItems(hearthstoneToys)
-	LoadItems(wormholes)
+	LoadItems(tpm.Wormholes)
+	LoadItems(tpm.Hearthstones)
+	LoadItems(tpm.ItemTeleports)
 
 	local allLoaded = true
 	local function OnItemsLoaded()
 		if allLoaded then
 			tpm:Setup()
 			tpm:LoadOptions()
+			self:UnregisterEvent("ADDON_LOADED")
 		else
 			checkItemsLoaded(self)
 		end
@@ -1027,18 +914,24 @@ local function checkItemsLoaded(self)
 end
 
 function tpm:Setup()
-	if db.iconSize then
-		globalWidth = db.iconSize
-		globalHeight = db.iconSize
+	if db["Button:Size"] then
+		globalWidth = db["Button:Size"]
+		globalHeight = db["Button:Size"]
 	end
 
-	tpm:updateAvailableHearthstones()
-	tpm:updateAvailableWormholes()
-	tpm:updateAvailableSeasonalTeleport()
+	tpm:UpdateAvailableHearthstones()
+	tpm:UpdateAvailableWormholes()
+	tpm:UpdateAvailableSeasonalTeleports()
+	tpm:UpdateAvailableItemTeleports()
 
-	if db.hearthstone and db.hearthstone ~= "rng" and db.hearthstone ~= "none" and not PlayerHasToy(db.hearthstone) then
-		print(APPEND .. L["Hearthone Reset Error"]:format(db.hearthstone))
-		db.hearthstone = "none"
+	if
+		db["Teleports:Hearthstone"]
+		and db["Teleports:Hearthstone"] ~= "rng"
+		and db["Teleports:Hearthstone"] ~= "none"
+		and not PlayerHasToy(db["Teleports:Hearthstone"] --[[@as integer]])
+	then
+		print(APPEND .. L["Hearthone Reset Error"]:format(db["Teleports:Hearthstone"]))
+		db["Teleports:Hearthstone"] = "none"
 		tpm:updateHearthstone()
 	end
 
@@ -1046,19 +939,70 @@ function tpm:Setup()
 	hooksecurefunc("ToggleGameMenu", createAnchors)
 end
 
-local function OnEvent(self, event, addOnName)
-	if addOnName == "TeleportMenu" then
-		db = tpm:GetOptions()
-		db.debug = false
-	elseif event == "PLAYER_LOGIN" then
-		checkItemsLoaded(self)
-	end
-end
-
+-- Event Handlers
+local events = {}
+local normalizedSeasons = {
+	[13] = 1, -- TWW Season 1
+	[14] = 2, -- TWW Season 2
+}
 local f = CreateFrame("Frame")
 f:RegisterEvent("ADDON_LOADED")
 f:RegisterEvent("PLAYER_LOGIN")
-f:SetScript("OnEvent", OnEvent)
+f:RegisterEvent("BAG_UPDATE_DELAYED")
+f:RegisterEvent("CVAR_UPDATE")
+f:SetScript("OnEvent", function(self, event, ...)
+	events[event](self, ...)
+end)
+
+function events:ADDON_LOADED(...)
+	local addOnName = ...
+
+	if addOnName == "TeleportMenu" then
+		db = tpm:GetOptions()
+		tpm.settings.current_season = normalizedSeasons[tonumber(C_CVar.GetCVar("newMythicPlusSeason"))] or 1
+
+		db.debug = false
+		f:UnregisterEvent("ADDON_LOADED")
+	end
+end
+
+function events:CVAR_UPDATE(...)
+	local name, value = ...
+	if name == "newMythicPlusSeason" then
+		tpm.settings.current_season = normalizedSeasons[tonumber(value)] or 1
+		if TeleportMeButtonsFrame then
+			tpm:UpdateAvailableSeasonalTeleports()
+			tpm:ReloadFrames()
+		end
+	end
+end
+
+function events:PLAYER_LOGIN(...)
+	checkItemsLoaded(f)
+	f:UnregisterEvent("PLAYER_LOGIN")
+end
+
+function events:BAG_UPDATE_DELAYED()
+	--- @type Item[]
+	local items_in_possession = CopyTable(tpm.player.items_in_possession)
+
+	--- @type Item[]
+	local items_to_be_obtained = CopyTable(tpm.player.items_to_be_obtained)
+
+	-- Scan bags for items supposedly in possession
+	for _, item in pairs(items_in_possession) do
+		if GetItemCount(item.id) == 0 then
+			tpm:RemoveItemFromPossession(item.id)
+		end
+	end
+
+	-- Scan bags for items supposedly NOT in possession
+	for _, item in pairs(items_to_be_obtained) do
+		if GetItemCount(item.id) > 0 then
+			tpm:AddItemToPossession(item.id)
+		end
+	end
+end
 
 -- Debug Functions
 function tpm:DebugPrint(...)