+++ AddOns/TeleportMenu/CHANGELOG.md
@@ -1,7 +1,15 @@
 # Teleport Menu
 
-## [v9.5](https://github.com/Justw8/TeleportMenu/tree/v9.5) (2024-11-03)
-[Full Changelog](https://github.com/Justw8/TeleportMenu/compare/v9.4...v9.5) [Previous Releases](https://github.com/Justw8/TeleportMenu/releases)
+## [v10](https://github.com/Justw8/TeleportMenu/tree/v10) (2025-02-25)
+[Full Changelog](https://github.com/Justw8/TeleportMenu/compare/v9.5...v10) [Previous Releases](https://github.com/Justw8/TeleportMenu/releases)
 
-- Update chat print when using the chat command  
-- Options: Properly use defaults  
+- TOC update for 11.1 (#86)  
+- Add Vulpera Racials & Druid's Moonglade teleport (#85)  
+    Co-authored-by: Mike Donkers <67101689+donkers-mike@users.noreply.github.com>  
+- Remove dependency on lod blizzard data (#84)  
+- Update zhTW.lua (#83)  
+    Co-authored-by: Matthijs Groot // Justwait <Justw8@users.noreply.github.com>  
+- Update zhCN.lua (#82)  
+- Add item teleports (#81)  
+- Implement season 2 teleports (#79)  
+- Update zhCN.lua (#72)